Use 64-bit compatible index for operator[].

+++ vec3.h
@@ -20,8 +20,8 @@ public:
 
     inline const Vec3& operator+() const { return *this; }
     inline Vec3 operator-() const { return Vec3(-e[0], -e[1], -e[2]); }
-    inline float operator[](int i) const { return e[i]; }
-    inline float& operator[](int i) { return e[i]; }
+    inline float operator[](size_t i) const { return e[i]; }
+    inline float& operator[](size_t i) { return e[i]; }
 
     inline Vec3& operator+=(const Vec3& o);
     inline Vec3& operator-=(const Vec3& o);
